requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
google-generativeai>=0.5.2
python-dotenv==1.0.1
